build: move views from recognizer to personal account
