+++ EmotionRecognizer/personalaccountapp/urls.py
@@ -1,0 +1,9 @@
+from django.urls import path
+from .views import *
+
+
+urlpatterns = [
+    path('', LoginUser.as_view(), name='login'),
+    path('register/', RegisterUser.as_view(), name='register'),
+    path('settings/', change_password, name='user_settings'),
+]
